requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
chmod==0.0.8
click==8.1.7
colorama==0.4.6
customtkinter==5.2.2
darkdetect==0.8.0
decorator==5.1.1
distro==1.9.0
editdistpy==0.1.5
et_xmlfile==2.0.0
Faker==29.0.0
Flask==3.0.3
frozenlist==1.5.0
h11==0.14.0
html-testRunner==1.2.1
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
imap-tools==1.7.4
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.9.0
MarkupSafe==3.0.1
moviepy==2.1.1
multidict==6.2.0
nodemon==0.0.16
numpy==2.1.3
openai>=2.0.0
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
parameterized==0.9.0
pefile==2023.2.7
pillow==10.4.0
pluggy==1.5.0
proglog==0.1.10
propcache==0.3.1
pycparser==2.22
pydantic==2.11.1
pydantic_core==2.33.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.11
PySocks==1.7.1
pytest==8.3.3
pytest-html==4.1.1
pytest-metadata==3.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32-ctypes==0.2.3
requests==2.32.3
selenium==4.27.0
setuptools==75.6.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
source==1.2.0
symspellpy==6.9.0
tqdm==4.67.1
trio==0.26.2
trio-websocket==0.11.1
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.0.4
wget==3.2
wheel==0.44.0
wsl==0.12
wsproto==1.2.0
yarl==1.18.3
